Setting default timeout of 20s for all sockets
I experienced a loss of connectivity today, and rather than failing or
throwing an exception the bot just kind of hung and failed to do anything
further. When this happened before, it was because the twitch API socket
was not ever timing out.

I suspect there is a similar issue going on here, but perhaps in one of
libraries used by the bot. I hope that setting the global timeout will
solve this issue and at least bring forth an exception to handle if not
resolve the issue entirely.